requirements.txt
google-cloud-vision
google-cloud-speech
pydub
google-generativeai
python-dotenv
deep-translator
Pillow
fastapi
uvicorn
PyMuPDF
